+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/ClientE2ETestSuite.scala
@@ -17,7 +17,6 @@
 package org.apache.spark.sql
 
 import java.io.{ByteArrayOutputStream, PrintStream}
-import java.lang.ref.WeakReference
 import java.nio.file.Files
 import java.time.DateTimeException
 import java.util.Properties
@@ -31,12 +30,9 @@ import org.apache.commons.io.FileUtils
 import org.apache.commons.io.output.TeeOutputStream
 import org.scalactic.TolerantNumerics
 import org.scalatest.PrivateMethodTester
-import org.scalatest.concurrent.Eventually.{eventually, interval, timeout}
-import org.scalatest.exceptions.TestFailedDueToTimeoutException
 
 import org.apache.spark.{SparkArithmeticException, SparkException, SparkUpgradeException}
 import org.apache.spark.SparkBuildInfo.{spark_version => SPARK_VERSION}
-import org.apache.spark.connect.proto
 import org.apache.spark.sql.catalyst.analysis.{NamespaceAlreadyExistsException, NoSuchDatabaseException, TableAlreadyExistsException, TempTableAlreadyExistsException}
 import org.apache.spark.sql.catalyst.encoders.AgnosticEncoders.StringEncoder
 import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema
@@ -1561,75 +1557,6 @@ class ClientE2ETestSuite
     // make sure the thread is unblocked after the query is finished
     val metrics = SparkThreadUtils.awaitResult(future, 2.seconds)
     assert(metrics === Map("min(id)" -> 0, "avg(id)" -> 49, "max(id)" -> 98))
-  }
-
-  test("checkpoint") {
-    val df = spark.range(100).localCheckpoint()
-    testCapturedStdOut(df.explain(), "ExistingRDD")
-  }
-
-  test("checkpoint gc") {
-    var df1 = spark.range(100).localCheckpoint(eager = true)
-    val encoder = df1.agnosticEncoder
-    val dfId = df1.plan.getRoot.getCachedRemoteRelation.getRelationId
-
-    // GC triggers remove the cached remote relation
-    df1 = null
-    val ref = new WeakReference[Object](df1)
-    while (ref.get() != null) { Thread.sleep(1000L); System.gc() }
-
-    eventually(timeout(60.seconds), interval(1.second)) {
-      val ex = intercept[SparkException] {
-        spark
-          .newDataset(encoder) { builder =>
-            builder.setCachedRemoteRelation(
-              proto.CachedRemoteRelation
-                .newBuilder()
-                .setRelationId(dfId)
-                .build())
-          }
-          .collect()
-      }
-      assert(ex.getMessage.contains(s"No DataFrame with id $dfId is found"))
-    }
-  }
-
-  test("checkpoint gc derived DataFrame") {
-    var df1 = spark.range(100).localCheckpoint(eager = true)
-    var derived = df1.repartition(10)
-    val encoder = df1.agnosticEncoder
-    val dfId = df1.plan.getRoot.getCachedRemoteRelation.getRelationId
-
-    df1 = null
-    val ref = new WeakReference[Object](df1)
-    while (ref.get() != null) { Thread.sleep(1000L); System.gc() }
-
-    def condition(): Unit = {
-      val ex = intercept[SparkException] {
-        spark
-          .newDataset(encoder) { builder =>
-            builder.setCachedRemoteRelation(
-              proto.CachedRemoteRelation
-                .newBuilder()
-                .setRelationId(dfId)
-                .build())
-          }
-          .collect()
-      }
-      assert(ex.getMessage.contains(s"No DataFrame with id $dfId is found"))
-    }
-
-    intercept[TestFailedDueToTimeoutException] {
-      eventually(timeout(5.seconds), interval(1.second))(condition())
-    }
-
-    // GC triggers remove the cached remote relation
-    derived = null
-    val ref1 = new WeakReference[Object](df1)
-    while (ref1.get() != null) { Thread.sleep(1000L); System.gc() }
-
-    // Check the state was removed up on garbage-collection.
-    eventually(timeout(60.seconds), interval(1.second))(condition())
   }
 }
